Move RealTorManagerAndroid to internal file prefixed with -
